backend/requirements.txt
pandas
faiss-cpu
scikit-learn
sentence-transformers
fastapi
uvicorn
openai==0.28